fruits_lib.py
import os
import glob
import random

import numpy as np
from PIL import Image
import matplotlib.pyplot as plt
from matplotlib.animation import FuncAnimation
import os
import torch
import torch.nn as nn
import torch.optim as optim
import torch.utils.data as data
import torchvision
from torchvision import models, transforms
import matplotlib.pyplot as plt

torch.manual_seed(1009)
np.random.seed(1009)
random.seed(1009)

