requirements.txt
# Python 3.10
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
pillow==10.4.0
numpy==1.26.0
matplotlib==3.9.1

